refactor: separate plugin folders
