+++ src/components/Editor/plugins/History/index.tsx
@@ -1,5 +1,5 @@
 import { HistoryPlugin as LexicalHistoryPlugin } from '@lexical/react/LexicalHistoryPlugin';
-import { useEditorHistoryContext } from '../context/useEditorHistoryContext';
+import { useEditorHistoryContext } from '../../context/useEditorHistoryContext';
 
 export function HistoryPlugin() {
   const { historyState } = useEditorHistoryContext();
