+++ src/components/Editor/plugins/MarkdownShortcut/index.tsx
@@ -1,6 +1,6 @@
 import { MarkdownShortcutPlugin as LexicalMarkdownShortcutPlugin } from '@lexical/react/LexicalMarkdownShortcutPlugin';
 import { TRANSFORMERS, CODE, INLINE_CODE } from '@lexical/markdown';
-import { removeElements } from '../../../utils/removeElements';
+import { removeElements } from '../../../../utils/removeElements';
 
 const NO_USE = [CODE, INLINE_CODE];
 const CUSTOM_TRANSFORMERS = removeElements(TRANSFORMERS, NO_USE);
